requirements.txt
gradio==4.42.0
fastapi==0.112.2
pytest==8.3.2
wheel==0.44.0
packaging==24.1
torch
numpy==1.24.4
Pillow==10.3.0
Requests==2.31.0
torchvision==0.18.0
transformers==4.43.0
accelerate==0.30.0
diffusers==0.30.2
scipy==1.14.1
safetensors==0.4.5